tests

+++ fence_token_vending.py
@@ -19,8 +19,13 @@ class FenceTokenVendingMachine:
 
     def remove_service_account(self, user_id):
         fsa_key = ndb.Key(FenceServiceAccount, user_id)
-        fsa_key.delete()
-        memcache.delete(namespace='fence_key', key=user_id)
+        fence_service_account = fsa_key.get()
+        if fence_service_account:
+            access_token = self._get_oauth_access_token(user_id)
+            key_id = json.loads(fence_service_account.key_json)["private_key_id"]
+            # deleting the key will invalidate anything cached
+            self.fence_api.delete_credentials_google(access_token, key_id)
+            fsa_key.delete()
 
     def get_service_account_access_token(self, user_info, scopes=None):
         """
@@ -80,7 +85,7 @@ class FenceTokenVendingMachine:
         :return:
         """
         # get access_token before acquiring lock to keep lock duration as small as possible
-        access_token = self._get_oauth_access_token(real_user_info)
+        access_token = self._get_oauth_access_token(real_user_info[SamKeys.USER_ID_KEY])
 
         if self._acquire_lock(fsa_key):
             key_json = self.fence_api.get_credentials_google(access_token)
@@ -98,8 +103,8 @@ class FenceTokenVendingMachine:
 
         return fence_service_account
 
-    def _get_oauth_access_token(self, real_user_info):
-        refresh_token = TokenStore.lookup(real_user_info["userSubjectId"])
+    def _get_oauth_access_token(self, user_id):
+        refresh_token = TokenStore.lookup(user_id)
         if refresh_token is None:
             raise endpoints.BadRequestException("Fence account not linked")
         access_token = self.fence_oauth_adapter.refresh_access_token(refresh_token.token).get("access_token")
